pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<4.0"
dependencies = [
    "elevenlabs>=2.30.0",
    "gradio>=6.3.0",
    "groq>=0.33.0",
    "gtts>=2.5.4",
    "langchain>=1.2.4",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.6",
    "opencv-python>=4.12.0.88",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "speechrecognition>=3.14.5",
]
